+++ internal/types/gatewaytypes_interfaces.go
@@ -4,7 +4,7 @@ import "sigs.k8s.io/controller-runtime/pkg/client"
 
 // SupportedRoute defines a supported route type.
 type SupportedRoute interface {
-	HTTPRoute | TLSRoute | GRPCRoute | UDPRoute
+	HTTPRoute | TLSRoute | TCPRoute | GRPCRoute | UDPRoute
 }
 
 // SupportedRoutePtr defines a pointer of a supported route type.
@@ -16,7 +16,7 @@ type SupportedRoutePtr[T SupportedRoute] interface {
 
 // SupportedRouteList defines a list of supported route.
 type SupportedRouteList interface {
-	HTTPRouteList | TLSRouteList | UDPRouteList
+	HTTPRouteList | TLSRouteList | TCPRouteList | UDPRouteList
 }
 
 // SupportedRouteListPtr defines a pointer of a supported route list.
@@ -28,7 +28,7 @@ type SupportedRouteListPtr[T SupportedRouteList] interface {
 
 // SupportedRouteRule defines a rule in a supported route.
 type SupportedRouteRule interface {
-	HTTPRouteRule | TLSRouteRule | UDPRouteRule
+	HTTPRouteRule | TLSRouteRule | TCPRouteRule | UDPRouteRule
 }
 
 // SupportedBackendRef defines a supported backendRef type.
@@ -42,6 +42,8 @@ func GetSpecParentRefs[T SupportedRoute](route T) []ParentReference {
 	case HTTPRoute:
 		return r.Spec.ParentRefs
 	case TLSRoute:
+		return r.Spec.ParentRefs
+	case TCPRoute:
 		return r.Spec.ParentRefs
 	case GRPCRoute:
 		return r.Spec.ParentRefs
@@ -59,6 +61,8 @@ func GetSpecHostnames[T SupportedRoute](route T) []Hostname {
 		return r.Spec.Hostnames
 	case TLSRoute:
 		return r.Spec.Hostnames
+	case TCPRoute:
+		return []Hostname{}
 	}
 	return []Hostname{}
 }